+++ sites/fast-component-explorer/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @microsoft/fast-component-explorer
 
-This log was last generated on Sun, 04 Jul 2021 07:15:28 GMT and should not be manually modified.
+This log was last generated on Wed, 07 Jul 2021 07:19:51 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.17.2
+
+Wed, 07 Jul 2021 07:19:51 GMT
+
+### Patches
+
+- Bump @microsoft/fast-tooling-react to v2.8.25 (scomea@microsoft.com)
 
 ## 0.17.2